requirements.txt
streamlit>=1.28
pandas>=1.5
openpyxl>=3.1
groq>=0.4
python-dotenv>=1.0